LPS-167572 SF, move logic from Scanner to DirectoryWatcher, prepare for next
