+++ modules/apps/static/portal-file-install/portal-file-install-impl/src/main/java/com/liferay/portal/file/install/internal/Scanner.java
@@ -17,7 +17,6 @@ package com.liferay.portal.file.install.internal;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.Validator;
 
 import java.io.File;
 import java.io.FilenameFilter;
@@ -37,8 +36,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.zip.CRC32;
 
 /**
@@ -49,27 +46,11 @@ public class Scanner {
 	public static final String SUBDIR_MODE_RECURSE = "recurse";
 
 	public Scanner(
-		List<File> dirs, final String filterString, String subdirMode) {
+		List<File> dirs, FilenameFilter filenameFilter, String subdirMode) {
+
+		_filenameFilter = filenameFilter;
 
 		_watchedDirs = _canononize(dirs);
-
-		if (!Validator.isBlank(filterString)) {
-			_filenameFilter = new FilenameFilter() {
-
-				@Override
-				public boolean accept(File dir, String name) {
-					Matcher matcher = _pattern.matcher(name);
-
-					return matcher.matches();
-				}
-
-				private final Pattern _pattern = Pattern.compile(filterString);
-
-			};
-		}
-		else {
-			_filenameFilter = (dir, name) -> true;
-		}
 
 		_recurseSubdir = SUBDIR_MODE_RECURSE.equals(subdirMode);
 	}
